+++ src/Setups/Widgets/AbstractWidget.php
@@ -1,0 +1,10 @@
+<?php
+declare (strict_types = 1);
+
+namespace GrottoPress\WordPress\SUV\Setups\Widgets;
+
+use GrottoPress\WordPress\SUV\Setups\AbstractSetup;
+
+abstract class AbstractWidget extends AbstractSetup
+{
+}