+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Add shortcodes setup
+- Add widgets setup
 
 ## [1.0.0] - 2023-05-31